refactor(backend): isolate direct visible thinking prompts
- Move visible-thinking supplement and domain example loading into direct_prompt_visible_thinking.py

- Keep direct_prompts.py as the assembly shell while preserving prompt text

- Update runtime cleanup audit with the #535 boundary

Refs #535
